// frontend/src/app/models/user.ts
export class User {
    ime: String;
    prezime: String;
    username: String;
    password : String;
    datumRodj: String;
    grad: String;
    drzava: String;
    email: String;
    slika: String;
    tip: String;
    pratioci: Array<String>;
    obavestenja: Array<String>;
    aktivan: String;
}